rename profileUtils
