+++ src/utils/profileUtils.ts
